added(http, http_test): renamed URL to HOST
